// testservices/models/ArtifactCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArtifactCreate = {
    artifact_type: string;
    url: string;
    download_method: string;
    version: string;
    checksum: string;
    labels?: (Array<string> | null);
    association_target?: (string | null);
};