// testservices/models/ArtifactUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArtifactUpdate = {
    artifact_type?: (string | null);
    url?: (string | null);
    download_method?: (string | null);
    version?: (string | null);
    checksum?: (string | null);
    labels?: (Array<string> | null);
};