// testservices/models/LimitOffsetPage_ArtifactRead_.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ArtifactRead } from './ArtifactRead';
export type LimitOffsetPage_ArtifactRead_ = {
    items: Array<ArtifactRead>;
    total: (number | null);
    limit: (number | null);
    offset: (number | null);
};